Move Emscripten-specific bits of WebGPU out from html5.h to a dedicated emscripten_webgpu.h header and JS library file
